+++ pyhetdex/coordinates/astrometry.py
@@ -44,9 +44,9 @@ _group_astro.add_argument('--astrometry', nargs=3, type=float64,
                           (degrees)''')
 _group_astro.add_argument('--image', help='''An image, with a header to grab
                           ra, dec and PA from (DONT USE THIS)''')
-astro_parent.add_argument('--rhozp', default=1.3, type=float,
-                          help='''Set the rho zeropoint in degrees, the
-                          rotation of the IHMP wrt HET (default=1.3 deg)''')
+astro_parent.add_argument('--rhozp', default=0.0, type=float,
+                          help='''Optionally set the rho zeropoint in degrees, i.e. the
+                          rotation of the IHMP wrt HET (default=0.0 deg)''')
 
 
 def ihmp_astrometry(opts, xscale=1.0, yscale=1.0):